Rename message tests to violation tests
